+++ CHANGES.rst
@@ -7,7 +7,9 @@ Changelog
 Features
 ********
 
-- PR #112: Add IPv6 support. Patch by Dan Kruchinin.
+- #115: Limit the backends we use for SLF4J during tests.
+
+- #112: Add IPv6 support. Patch by Dan Kruchinin.
 
 1.2.1 (2013-08-01)
 ------------------